backend/requirements.txt
Django>=5.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9

# Testing & Linting
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
mixer>=7.2.2
ruff>=0.3.0
